+++ tests/fixtures/token_capturing.rs
@@ -106,6 +106,7 @@ pub fn parse(
         encoding,
         next_encoding: Default::default(),
         strict: true,
+        graceful_bail_out_on_memory_limit_exceeded: false,
     });
 
     let parser = transform_stream.parser();